requirements.txt
streamlit
yfinance
statsmodels
vaderSentiment
plotly.graph_objs
pandas
numpy